+++ crates/vertigo-cli/src/serve/html/element_children.rs
@@ -7,31 +7,33 @@ struct ChildrenNode {
 
 pub struct ElementChildren {
     first_child: Option<u64>,
-    childs: HashMap<u64, ChildrenNode>,
+    children: HashMap<u64, ChildrenNode>,
 }
 
 impl ElementChildren {
     pub fn new() -> ElementChildren {
         ElementChildren {
             first_child: None,
-            childs: HashMap::new(),
+            children: HashMap::new(),
         }
     }
 
-    pub fn childs(&self) -> Vec<u64> {
-        if self.childs.is_empty() {
+    pub fn get_all(&self) -> Vec<u64> {
+        if self.children.is_empty() {
             return Vec::new();
         }
 
         let mut result = Vec::new();
 
         let Some(mut wsk_id) = self.first_child else {
-            unreachable!();
+            log::error!("Unreachable in ElementChildren::get_all (1)");
+            return Vec::new();
         };
 
-        for _ in 0..self.childs.len() {
-            let Some(node) = self.childs.get(&wsk_id) else {
-                unreachable!();
+        for _ in 0..self.children.len() {
+            let Some(node) = self.children.get(&wsk_id) else {
+                log::error!("Unreachable in ElementChildren::get_all (2)");
+                return result;
             };
 
             result.push(wsk_id);
@@ -45,19 +47,21 @@ impl ElementChildren {
     }
 
     fn insert_on_left(&mut self, node_right_id: u64, node_id: u64) {
-        let Some(node_right) = self.childs.get_mut(&node_right_id) else {
-            unreachable!();
+        let Some(node_right) = self.children.get_mut(&node_right_id) else {
+            log::error!("Unreachable in ElementChildren::insert_on_left (1)");
+            return;
         };
 
         let node_left_id = node_right.left;
         node_right.left = node_id;
 
-        let Some(node_left) = self.childs.get_mut(&node_left_id) else {
-            unreachable!();
+        let Some(node_left) = self.children.get_mut(&node_left_id) else {
+            log::error!("Unreachable in ElementChildren::insert_on_left (2)");
+            return;
         };
         node_left.right = node_id;
 
-        self.childs.insert(
+        self.children.insert(
             node_id,
             ChildrenNode {
                 left: node_left_id,
@@ -68,7 +72,8 @@ impl ElementChildren {
 
     pub fn insert_before(&mut self, ref_id: Option<u64>, child: u64) {
         if ref_id == Some(child) {
-            panic!("ref_id must not be equal to child_id, ref_id={ref_id:?}, child={child:?}");
+            log::error!("ref_id must not be equal to child_id, ref_id={ref_id:?}, child={child:?}");
+            return;
         }
 
         self.remove(child);
@@ -79,10 +84,10 @@ impl ElementChildren {
             } else if let Some(first_child) = self.first_child {
                 first_child
             } else {
-                assert_eq!(self.childs.len(), 0);
+                assert_eq!(self.children.len(), 0);
 
                 self.first_child = Some(child);
-                self.childs.insert(
+                self.children.insert(
                     child,
                     ChildrenNode {
                         left: child,
@@ -101,7 +106,7 @@ impl ElementChildren {
     }
 
     pub fn remove(&mut self, node_id: u64) {
-        let Some(node) = self.childs.remove(&node_id) else {
+        let Some(node) = self.children.remove(&node_id) else {
             return;
         };
 
@@ -112,14 +117,16 @@ impl ElementChildren {
             return;
         }
 
-        let Some(node_right) = self.childs.get_mut(&node.right) else {
-            unreachable!();
+        let Some(node_right) = self.children.get_mut(&node.right) else {
+            log::error!("Unreachable in ElementChildren::remove (1)");
+            return;
         };
 
         node_right.left = node.left;
 
-        let Some(node_left) = self.childs.get_mut(&node.left) else {
-            unreachable!();
+        let Some(node_left) = self.children.get_mut(&node.left) else {
+            log::error!("Unreachable in ElementChildren::remove (2)");
+            return;
         };
 
         node_left.right = node.right;
@@ -139,101 +146,101 @@ mod tests {
         let mut children = ElementChildren::new();
 
         assert_eq!(children.first_child, None);
-        assert_eq!(children.childs(), Vec::<u64>::new());
+        assert_eq!(children.get_all(), Vec::<u64>::new());
 
         children.insert_before(None, 44);
 
         assert_eq!(children.first_child, Some(44));
-        assert_eq!(children.childs(), vec!(44));
+        assert_eq!(children.get_all(), vec!(44));
 
         children.insert_before(None, 55);
 
         assert_eq!(children.first_child, Some(44));
-        assert_eq!(children.childs(), vec!(44, 55));
+        assert_eq!(children.get_all(), vec!(44, 55));
 
         children.insert_before(None, 66);
 
         assert_eq!(children.first_child, Some(44));
-        assert_eq!(children.childs(), vec!(44, 55, 66));
+        assert_eq!(children.get_all(), vec!(44, 55, 66));
 
         children.insert_before(Some(44), 33);
 
         assert_eq!(children.first_child, Some(33));
-        assert_eq!(children.childs(), vec!(33, 44, 55, 66));
+        assert_eq!(children.get_all(), vec!(33, 44, 55, 66));
 
         children.insert_before(Some(44), 35);
 
         assert_eq!(children.first_child, Some(33));
-        assert_eq!(children.childs(), vec!(33, 35, 44, 55, 66));
+        assert_eq!(children.get_all(), vec!(33, 35, 44, 55, 66));
 
         children.remove(55);
 
         assert_eq!(children.first_child, Some(33));
-        assert_eq!(children.childs(), vec!(33, 35, 44, 66));
+        assert_eq!(children.get_all(), vec!(33, 35, 44, 66));
 
         children.remove(66);
 
         assert_eq!(children.first_child, Some(33));
-        assert_eq!(children.childs(), vec!(33, 35, 44));
+        assert_eq!(children.get_all(), vec!(33, 35, 44));
 
         children.remove(33);
 
         assert_eq!(children.first_child, Some(35));
-        assert_eq!(children.childs(), vec!(35, 44));
+        assert_eq!(children.get_all(), vec!(35, 44));
 
         children.insert_before(Some(44), 36);
 
         assert_eq!(children.first_child, Some(35));
-        assert_eq!(children.childs(), vec!(35, 36, 44));
+        assert_eq!(children.get_all(), vec!(35, 36, 44));
 
         children.remove(35);
 
         assert_eq!(children.first_child, Some(36));
-        assert_eq!(children.childs(), vec!(36, 44));
+        assert_eq!(children.get_all(), vec!(36, 44));
 
         children.remove(36);
 
         assert_eq!(children.first_child, Some(44));
-        assert_eq!(children.childs(), vec!(44));
+        assert_eq!(children.get_all(), vec!(44));
 
         children.remove(44);
 
         assert_eq!(children.first_child, None);
-        assert_eq!(children.childs(), Vec::<u64>::new());
+        assert_eq!(children.get_all(), Vec::<u64>::new());
 
         children.insert_before(None, 9999);
 
         assert_eq!(children.first_child, Some(9999));
-        assert_eq!(children.childs(), vec!(9999));
+        assert_eq!(children.get_all(), vec!(9999));
 
         children.insert_before(None, 9999);
 
         assert_eq!(children.first_child, Some(9999));
-        assert_eq!(children.childs(), vec!(9999));
+        assert_eq!(children.get_all(), vec!(9999));
 
         children.insert_before(Some(9999), 8888);
 
         assert_eq!(children.first_child, Some(8888));
-        assert_eq!(children.childs(), vec!(8888, 9999));
+        assert_eq!(children.get_all(), vec!(8888, 9999));
 
         children.insert_before(Some(9999), 8888);
 
         assert_eq!(children.first_child, Some(8888));
-        assert_eq!(children.childs(), vec!(8888, 9999));
+        assert_eq!(children.get_all(), vec!(8888, 9999));
 
         children.insert_before(Some(9999), 8900);
 
         assert_eq!(children.first_child, Some(8888));
-        assert_eq!(children.childs(), vec!(8888, 8900, 9999));
+        assert_eq!(children.get_all(), vec!(8888, 8900, 9999));
 
         children.insert_before(Some(9999), 8900);
 
         assert_eq!(children.first_child, Some(8888));
-        assert_eq!(children.childs(), vec!(8888, 8900, 9999));
+        assert_eq!(children.get_all(), vec!(8888, 8900, 9999));
 
         children.insert_before(Some(8900), 9999);
 
         assert_eq!(children.first_child, Some(8888));
-        assert_eq!(children.childs(), vec!(8888, 9999, 8900));
+        assert_eq!(children.get_all(), vec!(8888, 9999, 8900));
     }
 }